+++ src/cartas-per/cartas-per.service.ts
@@ -289,6 +289,11 @@ export class CartasPerService {
       .innerJoin('carta.estudiante', 'alumno', 'alumno.run = :run', { run })
       .select('carta.id')
       .getRawMany();
+
+    if (cartas.length == 0) {
+      return 'No existe la carta';
+    }
+    
     return await this.CartasPerRepository.delete(cartas[0].carta_id);
   }
 }
